Added Argument Parsing & Some Other Changes
Added basic input pasing for the command line. Now users should be easily able to specify from which file to read and to which folder to write. (Have to add the option to choose between export formats)

Video files are removed after download. (Have to figure out how to load it into memory instead for conversion to not take up hard drive space)

Added pytube dependancy to setup.py. Version 10.1.0, or higher, is needed for the changes made in the 'song_downloaded.py' script.